+++ sei-ibc-go/modules/apps/27-interchain-accounts/controller/client/cli/cli.go
@@ -14,6 +14,7 @@ func GetQueryCmd() *cobra.Command {
 	}
 
 	queryCmd.AddCommand(
+		GetCmdQueryInterchainAccount(),
 		GetCmdParams(),
 	)