+++ sei-ibc-go/modules/apps/27-interchain-accounts/controller/keeper/grpc_query.go
@@ -4,16 +4,41 @@ import (
 	"context"
 
 	sdk "github.com/cosmos/cosmos-sdk/types"
+	"google.golang.org/grpc/codes"
+	"google.golang.org/grpc/status"
 
 	"github.com/cosmos/ibc-go/v3/modules/apps/27-interchain-accounts/controller/types"
+	icatypes "github.com/cosmos/ibc-go/v3/modules/apps/27-interchain-accounts/types"
 )
 
 var _ types.QueryServer = Keeper{}
 
+// InterchainAccount implements the Query/InterchainAccount gRPC method
+func (k Keeper) InterchainAccount(goCtx context.Context, req *types.QueryInterchainAccountRequest) (*types.QueryInterchainAccountResponse, error) {
+	if req == nil {
+		return nil, status.Error(codes.InvalidArgument, "empty request")
+	}
+
+	portID, err := icatypes.NewControllerPortID(req.Owner)
+	if err != nil {
+		return nil, status.Errorf(codes.InvalidArgument, "failed to generate portID from owner address: %s", err)
+	}
+
+	ctx := sdk.UnwrapSDKContext(goCtx)
+	addr, found := k.GetInterchainAccountAddress(ctx, req.ConnectionId, portID)
+	if !found {
+		return nil, status.Errorf(codes.NotFound, "failed to retrieve account address for %s on connection %s", portID, req.ConnectionId)
+	}
+
+	return &types.QueryInterchainAccountResponse{
+		Address: addr,
+	}, nil
+}
+
 // Params implements the Query/Params gRPC method
-func (q Keeper) Params(c context.Context, _ *types.QueryParamsRequest) (*types.QueryParamsResponse, error) {
+func (k Keeper) Params(c context.Context, _ *types.QueryParamsRequest) (*types.QueryParamsResponse, error) {
 	ctx := sdk.UnwrapSDKContext(c)
-	params := q.GetParams(ctx)
+	params := k.GetParams(ctx)
 
 	return &types.QueryParamsResponse{
 		Params: &params,